% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMismatches.R
\name{exportMismatches}
\alias{exportMismatches}
\title{Function to extract positions of mismatched bases from the BAM file with respect to the reference.}
\usage{
exportMismatches(
  bamfile = NULL,
  bamindex = bamfile,
  secondary.aln = FALSE,
  keep.duplicates = FALSE,
  region = NULL,
  min.mapq = 10,
  min.baseq = 20,
  min.mismatches = 5,
  by.phase = FALSE,
  ref.fasta = NULL,
  bsgenome = NULL
)
}
\arguments{
\item{bamfile}{A path to a BAM file.}

\item{min.baseq}{A minimum base quality per aligned base to be retained.}

\item{min.mismatches}{A minimum number of mismatch coverage to be reported.}

\item{by.phase}{If set to \code{TRUE} mismatches will be reported per haplotype specific BAM tag ('HP').}

\item{ref.fasta}{A path to a reference FASTA file to extract reference bases from.}

\item{bsgenome}{A \pkg{\link[BSgenome]{BSgenome-class}} object of reference genome to get reference bases from.}

\item{add.phase}{Set to \code{TRUE} if phasing specific tag 'HP' should be added if defined in the BAM file.}
}
\value{
A \code{list} containing two data frames objects called `mismatches` storing reported mismatches
and `coverage` reporting overall coverage per genomic position.
}
\description{
Function to extract positions of mismatched bases from the BAM file with respect to the reference.
}
\examples{
## Get BAM to extract mismatches from
bam.file <- system.file("extdata", "test2.bam", package = "ggNucFreak")
## Set region to extract mismatches from
region <- as('chr1:197373833-197391376', "GRanges")
## Load bsgnome object to obtain reference bases from
library(BSgenome.Hsapiens.UCSC.hg38)
## Extract mismatches
m.obj <- exportMismatches(bamfile = bam.file, region = region, bsgenome = BSgenome.Hsapiens.UCSC.hg38)
m.obj

}
\author{
David Porubsky
}
