% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMethylCs.R
\name{extractMethylCs}
\alias{extractMethylCs}
\title{Function to extract positions of methylated cytosines from the BAM file in forward reference coordinates.}
\usage{
extractMethylCs(
  bamfile = NULL,
  bamindex = bamfile,
  region = NULL,
  secondary.aln = FALSE,
  keep.duplicates = FALSE,
  min.mapq = 10,
  add.phase = FALSE
)
}
\arguments{
\item{bamfile}{A path to a BAM file containing methylation specific tags (MM and ML).}

\item{bamindex}{A path to the BAM file index. By default index file is expected to be present in the same directory. If not present it will be created.}

\item{region}{A user defined region to be extracted from the BAM.}

\item{secondary.aln}{Set to \code{TRUE} if secondary alignments should be retained.}

\item{keep.duplicates}{Set to \code{TRUE} if duplicate reads should be retained.}

\item{min.mapq}{A minimum mapping quality of aligned reads to be extracted from provided BAM file.}

\item{add.phase}{Set to \code{TRUE} if phasing specific tag 'HP' should be added if defined in the BAM file.}
}
\value{
A \code{list} of \code{\link{GRanges-class}} objects called `reads` storing read boundaries and positions
and `methylCs` reporting positions of methylated cytosines in reference coordinates.
}
\description{
Function to extract positions of methylated cytosines from the BAM file in forward reference coordinates.
}
\examples{
## Get BAM to extract methylation from
bam.file <- system.file("extdata", "test.bam", package = "ggNucFreak")
## Set region to extract methylation from
region <- as('chr11:2787889-2790647', "GRanges")
region <- GenomicRanges::resize(region, width = GenomicRanges::width(region) + (GenomicRanges::width(region) * 2), fix = 'center')
## Extract methylation
mCs <- extractMethylCs(bamfile = bam.file, region=region)
mCs

}
\author{
David Porubsky
}
