% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggNucFreq.R
\name{ggNucFreq}
\alias{ggNucFreq}
\title{Load BAM alignments a plot base frequencies.}
\usage{
ggNucFreq(
  bam.file = NULL,
  region = NULL,
  max.data.points = 1e+05,
  base.qual = 25,
  map.qual = 0,
  max.depth = 100,
  shape = "step",
  nucleotide.content = NULL,
  bin.size = 1e+05,
  reverse = FALSE,
  title = NULL
)
}
\arguments{
\item{bam.file}{...}

\item{region}{...}

\item{max.data.points}{...}

\item{base.qual}{...}

\item{map.qual}{...}

\item{max.depth}{...}

\item{shape}{...}

\item{nucleotide.content}{...}

\item{bin.size}{...}

\item{reverse}{Set to \code{TRUE}}

\item{title}{...}
}
\value{
A \code{ggplot} plots
}
\description{
This function takes loaded PAF alignments using \code{\link{readPaf}} function and perform
user defined filtering of input alignments based on mapping quality, alignment length, and
minimum alignments between target and query.
}
\author{
David Porubsky
}
