% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethyl.R
\name{plotMethyl}
\alias{plotMethyl}
\title{Function to plot methylated cytosines from the BAM file in forward reference coordinates.}
\usage{
plotMethyl(
  bamfile = NULL,
  region = NULL,
  min.mapq = 10,
  by.phase = FALSE,
  highlight.pos = NULL,
  highlight.gr = NULL
)
}
\arguments{
\item{bamfile}{A path to a BAM file containing methylation specific tags (MM and ML).}

\item{region}{A user defined region to be extracted from the BAM.}

\item{min.mapq}{A minimum mapping quality of aligned reads to be extracted from provided BAM file.}

\item{by.phase}{Set to \code{TRUE} if the methylation should be plotted separated by haplotypes.}

\item{highlight.pos}{A vector of x-axis coordinates to be highlighted by the a solid line.}

\item{highlight.gr}{A \code{\link{GRanges-class}} object of all regions to be highlighted by a transparent rectangle.}
}
\value{
A \code{\link{GRanges-class}} object with mirrored coordinates.
}
\description{
This function takes loaded methylation porfiles using \code{\link{extractMethylCs}} function and visualize them 
using basic \pkg{ggplot2} functions.
}
\examples{
## Get BAM file to plot
bam.file <- system.file("extdata", "test.bam", package = "ggNucFreak")
## Set region to plot methylation from
region <- as('chr11:2787889-2790647', "GRanges")
region.plus <- GenomicRanges::resize(region, width = GenomicRanges::width(region) + (GenomicRanges::width(region) * 2), fix = 'center')
## Plot methylation
plotMethyl(bamfile = bam.file, region = region.plus)
## Plot methylation by phase
plotMethyl(bamfile = bam.file, region = region.plus, by.phase = TRUE)
## Highlight specific positions
plotMethyl(bamfile = bam.file, region = region.plus, by.phase = TRUE, highlight.pos = c(2787889, 2790647))
## Highlight specific region
plotMethyl(bamfile = bam.file, region = region.plus, by.phase = TRUE, highlight.gr = region)

}
\author{
David Porubsky
}
