% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMismatches.R
\name{plotMismatches}
\alias{plotMismatches}
\title{Function to plot positions of mismatched bases extracted from the BAM file using \code{\link{exportMismatches}} function.}
\usage{
plotMismatches(
  mismatches.obj,
  add.reference.base = FALSE,
  coverage.profile = "line",
  highlight.pos = NULL,
  highlight.gr = NULL
)
}
\arguments{
\item{mismatches.obj}{\code{list} containing reported `mismatches` and `coverage` per genomic position.}

\item{add.reference.base}{Set to \code{TRUE} if reference base should be added to the plot.}

\item{coverage.profile}{Visulize overall coverage either as a solid 'line' or filled area 'fill'.}

\item{highlight.pos}{A vector of x-axis coordinates to be highlighted by the a solid line.}

\item{highlight.gr}{A \code{\link{GRanges-class}} object of all regions to be highlighted by a transparent rectangle.}
}
\value{
A \code{list} containing two data frames objects called `mismatches` storing reported mismatches
and `coverage` reporting overall coverage per genomic position.
}
\description{
Function to plot positions of mismatched bases extracted from the BAM file using \code{\link{exportMismatches}} function.
}
\examples{
## Get BAM to extract mismatches from
bam.file <- system.file("extdata", "test2.bam", package = "ggNucFreak")
## Set region to extract mismatches from
region <- as('chr1:197373833-197391376', "GRanges")
## Load bsgnome object to obtain reference bases from
library(BSgenome.Hsapiens.UCSC.hg38)
## Extract and plot mismatches
m.obj <- exportMismatches(bamfile = bam.file, region = region, bsgenome = BSgenome.Hsapiens.UCSC.hg38)
plotMismatches(mismatches.obj = m.obj)
## Extract and plot mismatches per haplotype
m.obj <- exportMismatches(bamfile = bam.file, region = region, bsgenome = BSgenome.Hsapiens.UCSC.hg38, by.phase=TRUE)
plotMismatches(mismatches.obj = m.obj)

}
\author{
David Porubsky
}
