% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNucFreq.R
\name{plotNucFreq}
\alias{plotNucFreq}
\title{Load BAM alignments a plot base frequencies.}
\usage{
plotNucFreq(
  bamfile = NULL,
  region = NULL,
  max.data.points = 1e+05,
  base.qual = 25,
  min.mapq = 0,
  max.depth = 100,
  shape = "point",
  nucleotide.content = NULL,
  bin.size = 1e+05,
  reverse = FALSE,
  title = NULL,
  return = "plot"
)
}
\arguments{
\item{bamfile}{A path to a BAM file to get base frequencies from.}

\item{region}{A user defined region from where the base frequencies will be extracted.}

\item{max.data.points}{A maximum number of data points to be plotted for efficiency (Default : `100000`).}

\item{base.qual}{A minimum base quality to be}

\item{min.mapq}{A minimum mapping quality of aligned reads to be extracted from provided BAM file.}

\item{max.depth}{(Default : `100`).}

\item{shape}{A user defined data point shape to visualize base frequency,either 'point' or 'step'.}

\item{nucleotide.content}{A user defined nucleotides whose frequency will be calculated, such 'GC' content.}

\item{bin.size}{A user defined}

\item{reverse}{Set to \code{TRUE}}

\item{title}{A user defined title of the final plot.}

\item{return}{Define to return plot only `plot`, plotting data only `data` or both `c('plot', 'data')` (Default : `plot`).}
}
\value{
A \code{ggplot} plot.
}
\description{
Load BAM alignments a plot base frequencies.
}
\examples{
## Get BAM to extract methylation from
bam.file <- system.file("extdata", "test.bam", package = "ggNucFreak")
## Set region to extract methylation from
region <- as('chr11:2787889-2790647', "GRanges")
region <- GenomicRanges::resize(region, width = GenomicRanges::width(region) + (GenomicRanges::width(region) * 2), fix = 'center')
## Plot frequencies of maximum and the remainder base(s).
plotNucFreq(bamfile = bam.file, region = region)
}
\author{
David Porubsky
}
